refactoring usage of getAge
